Plotly 'internal legend' shows 'legendString' value from tile's ts
